+++ src/assets/Close.icon.js
@@ -1,0 +1,16 @@
+/* eslint react/prop-types: 0 */
+import React from 'react';
+
+const CloseIcon = ({ className, onClick }) => (
+  <svg
+    className={className}
+    onClick={onClick}
+    width="24"
+    height="24"
+    viewBox="0 0 24 24"
+  >
+    <path d="M14.53 4.53l-1.06-1.06L9 7.94 4.53 3.47 3.47 4.53 7.94 9l-4.47 4.47 1.06 1.06L9 10.06l4.47 4.47 1.06-1.06L10.06 9z" />
+  </svg>
+);
+
+export default CloseIcon;
